410FunctionApp/requirements.txt
streamlit
python-dotenv
azure-functions